pyproject.toml
[tool.poetry]
name = "my-site"
version = "0.1.0"
description = "Static S3 web-site"
authors = ["Max Krivich <max@krivich.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-lib = "~=2.24"
constructs = "~=10.0.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
pytest = "^7.1.2"
black = "^22.3.0"
pylint = "^2.13.9"
flake8 = "^4.0.1"
mypy = "^0.950"
safety = "^1.10.3"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
